Add descriptive error when attempting to list tests before building
Typically a build is performed before a `swift test list` command, but
by passing the `--skip-build` flag it is possible to get a cryptic,
unhelpful erorr message if the test artifacts haven't already been
built.

Add a descriptive error for this case, only checking for test artifact
existence in the failure case to avoid any performance impact on the
happy path.