[WIP] Static versus exporting symbols for Windows
For Windows triples only, creates a parallel build graph for Swift
modules, one for static linking using -static, and one for
exporting linking which is the default. DLL products consume their
direct target dependencies as exporting. All other dependencies
use the static versions to eliminate unnecessary symbol exports.

The bulk of this is managed by the SwiftModuleBuildDescription
which will create a duplicate of itself for the exporting case
and set it's own type to static linking. Both modules are fed to
the planner to create llbuild swift command tasks. Code is added
for dynamic libraries to hook up the correct inputs for exporting
the symbols in the libraries targets.