`completion-tool list-executables` should list executable names (#8053)
Update the completion tool so it prints the list of executable names
instead of target names. The executable products names determines the
name fo the actual binary written to disk, which is what you'd want from
a completion-tool.

That patch also adds a few tests for completion-tool around this
specific behaviour. There weren't existing tests for the
completion-tool. More tests to follow in a subsequent PR focusing on
testing this area specifically.

Fixes #8033