Ensure plug-in C modules don't mix with target ones.
When a Swift module depends on a C library that is also a
dependent of plug-in that the module depends on, we were seeing
includes of the module map for both versions of that library,
host and target, causing types to be redefined.

This change filters out the plug-in version of these dependencies.