Support separate modules for static and dynamic exporting symbols for Windows (#8049)
On Windows, there is a limit of around 64K to the number of symbols a
DLL/EXE can export. We hit that regularly with large projects like
SwiftPM itself. This hides those symbols and only exposes the ones
requested for a DLL.

For Windows triples only, creates a parallel build graph for Swift
modules, one for static linking using -static, and one for exporting
linking which is the default. DLL products consume their direct target
dependencies as exporting. All other dependencies use the static
versions to eliminate unnecessary symbol exports.

The bulk of this is managed by the SwiftModuleBuildDescription which
will create a duplicate of itself for the exporting case and set its own
type to static linking. Both modules are fed to the planner to create
llbuild swift command tasks. Code is added for dynamic libraries to hook
up the correct inputs for exporting the symbols in the libraries
targets.

This is WIP as we need to do a lot of testing to ensure we didn't break
anything. Ensuring this only affects builds for Windows triples helps
mitigate that.