Expose all target files to SourceKit-LSP (#8107)
Add `resources`, `ignored`, and `other` files to `BuildTarget` for use
by SourceKit-LSP.

This is tracked by rdar://139431738 as part of
swiftlang/vscode-swift#562

### Motivation:

SourceKit-LSP needs to be able to determine which target contains a
given DocC catalog in order to fulfill documentation related requests.
However, the catalog is not listed as one of the `sources` or `headers`.
Rather, it is part of `other` files in most cases. Expose all file types
to SourceKit-LSP so that it can properly determine which target a given
file belongs to.

### Modifications:

Added three new properties to `BuildTarget`: `resources`, `ignored`, and
`other` which are arrays of `URL` pointing to the given file types in
the target.

### Result:

SourceKit-LSP will be able to see all files contained within a given
target.