Mark targets on Windows as static except those in DLLs.
Reverts the prevous attempt at fixing symbol export errors in EXEs.

Instead, for Windows triples, we mark all targets as needing to be
static. Then, after the products are planned, walks throught the
staticTargets of DLL products and marks as not static.

This allows executables in packages to import modules in local
DLL products as exporting. This is a temporary fix until we add
the ability for targets to explicity depend on local products.