Attempt two at solving Windows symbol linking issues (#8117)
Instead of creating two modules, one that exports symbols and one that
doesn't and letting consuming products pick which one they want, we now
mark all targets on Windows as static and then reset that if a target is
in a DLL product. This gets by the missing feature of marking a target
as a DLL (only products can be DLLs).

We also turn off the exporting imported symbol warning that is generated
when a module in a DLL is also used in a static context confusing the
linker.