Tests: Disable assert in TraitTests
The assertion in `TraitTets` that ensured warnigs do not appear in
stderr are failing against the 6.1 toolchain.  Since 1) this test was
never executed against macOS, since the self-hosted macOS pipeline was
running Swift 5.9.2, and 2) work is underway to productive the Traits
feature, it was decided to disable the assert for now, and open a GitHub
issue to re-enable the assert with the production work.

This change removes the assert for GitHub issue 8131, which must be
re-enabled when the respective issue is fixed.