Tests: Disable assert in TraitTests (#8132)
Disable assertion in `TraitTests`, which was never executed on macOS due
to the self-hosted macOS pipeline running Swift 5.9..2, and test being
wrapped around `#if compiler(>=6.0)` pre-processor.

### Motivation:

The assertion in `TraitTets` that ensured warnigs do not appear in
stderr are failing against the 6.1 toolchain. Since 1) this test was
never executed against macOS, since the self-hosted macOS pipeline was
running Swift 5.9.2, and 2) work is underway to productive the Traits
feature, it was decided to disable the assert for now, and open a GitHub
issue to re-enable the assert with the production work.

This change removes the assert for GitHub issue 8131, which must be
re-enabled when the respective issue is fixed.

### Modifications:

Disabled assertion the the `TraitTests`

### Result:

```
swift test --filter "TraitTests"
```


### NOTE:
It is expected the assertions will be re-introduced when #8131 is fixed.