Make GitShellError printable error.
### Motivation

Now and then `swift build` throws a GitShellError which doesn't have
details about what was an issue and with which dependency.

### Modification

GitShellError now conforms to CustomStringConvertible protocol which
makes it printable.

### Result

`swift build` now shows the error output from the git command
instead of showing an error code when GitShellError is thrown.