Do not throw an error from isValidDirectory if the directory is not a git repo.
### Motivation

isValidDirectory is called before a repo URL is established, and before a context
to throw a meaningful error is established. The errors that are thrown are not useful
to humans.

### Modification

This change supresses the GitShellError errors and allows to establish a context
for more meaningful GitCloneError errors.

### Result

Before the change:

> error: GitShellError:     fatal: cannot change to '/Users/yy/Library/Caches/org.swift.swiftpm/repositories/postgres-kit-e4358808': No such file or directory

After the change:

> error: Failed to clone repository https://github.com/vapor/postgres-kit.git:
>     Cloning into bare repository '/Users/yy/work/pub/tst/testpackage/.build/repositories/postgres-kit-e4358808'...
>     fatal: unable to access 'https://github.com/vapor/postgres-kit.git/': Could not resolve host: github.com