[cxx-interop] Do not pass `-lc++`/`-lstdc++` flags to the compiler
This makes sure that SwiftPM does not explicitly ask the compiler to link against a particular C++ stdlib.

The Swift compiler has recently introduced support for building with C++ stdlibs which aren't the default on the target platform, specifically libc++ on Linux. The Swift Driver has logic to determine which C++ stdlib should be used, depending on the target and explicit `-Xcc -stdlib=xyz` flags, and link against the requested standard library. Similar functionality exists in Clang Driver. We should defer the C++ stdlib selection to the compiler driver, it shouldn't be necessary to duplicate it in the build system.

rdar://141047307