Add `--experimental-attachments-path` to `swift test`.
This PR adds the given experimental command-line argument to `swift test` to allow writing
attachments to a user-supplied directory when using Swift Testing. In the future, we will
formally propose this flag as part of the attachments feature.

XCTest on Darwin would need separate changes to support this argument and they are beyond
the scope of this PR. XCTest on non-Darwin platforms (corelibs-xctest) does not support
attachments.

Resolves rdar://141225464.