Enable all traits when building the symbol graph
# Motivation

When building the symbol graph for documentation when want to make sure all trait guarded code is build.

# Modification

This PR changes the trait configuration when extracting the symbol graph by enabling all traits.

# Result

This PR will make sure all traits are enabled. In the future, we might want to pass a trait configuration through the DocC plugin to SwiftPM that allows us to build documentation with specific traits enabled but for now this should be fine since a package should support enabling all traits at once.