[Traits] Only allow valid Swift identifiers
# Motivation

During the proposal review it was brought up that the rules around trait names wasn't inline with Swift identifiers. Neither was the implementation which was inline with the proposal.

# Modification

This PR changes our validation logic to only allow valid Swift identifiers.

# Result

Only allows trait names that can be used as Swift identifiers.