pipeline: changes to support nightly toolchain (#8161)
Make changes to ensure success of the self-hosted macOS pipeline that
executes using the nightly toolchain.

### Motivation:

When enabling the nightly toolchain build, the CI build does not set
SDKROOT environment variable on Darwin, which causes some tests to fail

### Modifications:

1. Set SDKROOT in build-using-self when on Darwin. When using the
nightly toolchain, the `SDKROOT` environment variable was not set.
Explicitely set this to the default xcode installation only when on
Darwin.
2. Run the testTestsCanLinkAgainstAsyncExecutable only if it's executed
on macOS 15 or never.
3. Add additional logging to troubleshoot the bootstap failure and
convert to use python's logging module instead of print statements.

Verifies #8137