Move traits into resolution phase
Currently, target dependencies that are guarded by traits are being
calculated in the modules graph loading phase - there is a desire
to move this functionality into the resolution phase of SwiftPM to
avoid doing redundant calculations on dependencies that will not
be used further down the pipeline.

When a Workspace calls computeDependencies, there is a new set of
package identities being considered in the result that represent
the set of unused dependencies that are described in the root
package's set of dependencies - this assures that we can track
which manifests will be unnecessary to load, and also includes
dependencies that are guarded by traits that aren't enabled.

TODO:
* fully implement trait configurations passed along to manifest
* testing of the graph loading node correctness; see if dependencies
  that need to be omitted are in fact omitted by graph loading time
* as above, assure that unused dependencies aren't being downloaded as well
* refactor/replacement of extensions to appropriate files
* clean up comments/prints
* assure all paths to resolution consider traits