Add support for a prebuilt swift-syntax library for macros (#8142)
Following the pattern used for the early Traits work, added a workspace
prebuilts manager to co-ordinate detecting when macros are using
swift-syntax, downloading the zip file for the resolved swift-syntax
version, and extracting it into the scratch directory. We add the
necessary information about that to the workspace state. We then morph
the Modules that use it to add build settings to find the necessary
modules and libraries.

---------

Co-authored-by: Doug Schaefer <dschaefer2@gmail.com>