Fix prepare for indexing so plugin tools are properly built.
Eliminate the special generatePrepareManifest and just reuse the real
one and add a couple of checks for the target/product prepareForIndexing
buildParameter instead. This then enables us to make sure we're doing
a full build of the plugin tools so they can generate code to be
properly indexed.

Addresses #8216 for openapi generator.