Fetch Git LFS files after clone
Currently SPM clones repositories with `--mirror`, creating a bare
repository. If the cloned repository uses git lfs, the files in lfs are
never pulled and so the package resolution ultimately fails.

This is a naive implementation of LFS support that simply does a `lfs
fetch` on the mirrored repository, without regard for whether there are
LFS files to clone or even if the user has git lfs installed. It does,
however, allow packages using LFS to resolve correctly.

Issue: #8233