Set candidate URL for prebuilts download. (#8246)
I was using my fork of swift-syntax to host my test artifacts. Switch
that to a more realistic yet not yet approved one.

Also cuts down how often we reach out to fetch a manifest. We instead
create an empty one in the scratch directory if it's not found on the
download site. And we always now check the scratch directory first.