Enable building prebuilts on linux hosts without docker. (#8269)
Made discovery of the Linux distro reusable for the build-prebuilts
script. Changed it to use arch and os conditions instead of the host
toolchain.

Added support for Ubuntu noble, Fedora and Debian.

Removed the deletion of the stage directory so we can share it across
docker invocations for each distro. Fixed the Modules directory name
which wasn't working on case sensitive file systems.