Add a new `BuildDescription.load` API for use in SourceKit-LSP
Build tool plugins need to be built and run as part of the build plan
generation. Replace the existing API taking a `BuildPlan` with one that
generates it instead. This unfortunately exposes much more of the
SwiftPM API than previously, but given it is all required to a generate
a `BuildPlan`, this was really the case for all clients anyway.

Long term this module should be replaced by a long-running BSP server
that can prepare and provide build settings in the one process.

Resolves rdar://102242345.