Signing of the prebuilts manifest file. (#8284)
To ensure confidence that the manifest file and zip files are the ones
that were published, we use checksums for the zip files which are listed
in the manifest file. We now add signing for the manifest file to ensure
it isn't compromised. We use the same scheme as we used for package
collections where a signature is generated for the core manifest data
and then injected into the JSON file at the top level.

Note that we are adding the ManifestSigning to the Workspace module with
the eventual intention of making this a module we can reuse with package
collections once we forward port to main. We don't want to make changes
to the package collections on the 6.1 branch.

Also added a quick fix to the host platform computation that was causing
compile errors if not the two major ones (i.e. 32-bit Android ARM
hosts).