Add `package` access level to access level enums (#8261)
Currently the symbol graph generator defaults to "public" minimum access
level even if the user tries to initialize the options with "package"
this is because the internal enums representing access level don't have
a notion of `package` access control.

### Motivation:

Currently it is not possible to generate documentation only for
`package` and `public` symbols. This changes that

### Modifications:

Adds `package` values to enums representing access control along the
symbol graph generation code path.

### Result:

When invoking
[swift-docc-plugin](https://github.com/apple/swift-docc-plugin) with
`--symbol-graph-minimum-access-level package` the resulting
documentation contains `public` and `package` access controlled symbols.