Create an alternate SwiftPM build system called Swift Build
Add basic test coverage for the new build system for the supported platforms:
Linux, macOS, and Windows. Provide basic progress reporting, and an option
for "very verbose" mode so that that individual build tasks and the arguments
can be inspected for correctness.

Build a package with the new build system like this:

    swift build --build-system=swiftbuild

Enable very verbose task outputs with their flags, options, and arguments:

    swift build --build-system=swiftbuild --vv