Get run and test working against Swift Build.
Cleans up handling of the new build system. It's close to the old
xcode build system but with a couple of differences. Gets rid
of the boolean isXcodeBuild and pass in the actual Kind so we can
do different things based on which one is selected. Selects the
build output path for binaries based on the selection.

Also cleans up code with a missing wait for the build service to
complete and with with the file descriptor close before swift
run execs the process.