Move traits into resolution phase (#8205)
### Motivation:

Currently, target dependencies that are guarded by traits are being
calculated in the modules graph loading phase - there is a desire to
move this functionality into the resolution phase of SwiftPM to avoid
doing redundant calculations on dependencies that will not be used
further down the pipeline.

### Modifications:

Since there is an inherent reliance on expecting the resolution of
unused dependencies in SwiftPM's test suite, a flag
`--prune-unused-dependencies` has been added to act as a stopgap for
unexpected test failures and should only be considered as a feature for
trait-guarded target dependencies. This should be treated as an
experimental feature for any other unused dependencies for the time
being.

The `Manifest` is responsible for providing trait-related endpoints
since it stores all defined traits; there are new methods allowing for
the computation of which traits are enabled given a set of explicitly
enabled traits (either defined by the user passing in a trait
configuration through CLI, or defined by a manifest on its
`PackageDependency`), whether a trait guards a particular
`TargetDependency`, etc. The `dependenciesRequired` method considers
traits in its computation as well, depending on whether pruning
dependencies is enabled.

When a `Workspace` calls `computeDependencies`, there is a new set of
package identities being calculated in the result that represent the set
of unused dependencies (named `unusedDependencies`) that are described
in the root package's set of dependencies - this assures that we can
track which dependency manifests will be unnecessary to load. Note that
dependencies that are guarded by traits are inherently included in this
calculation.

A set of traits representing the **enabled traits** of a package have
been added as a property to both `PackageContainerConstraint` as well as
`DependencyResolutionNode`, since it is necessary to track this as a
requirement during dependency resolution and acts as a constraint on the
computation.

### Result:

With traits being factored into dependency resolution, SwiftPM can omit
guarded dependencies from its calculation thus avoiding unnecessary
checkouts and - subsequently - dependency version computations.