[Traits] Re-enable no unused dependencies assertion in `TraitTests.swift` (#8360)
Fixes #8131 

Previously this check was disabled since it introduced failures when
testing with the 6.1 toolchain - due to the introduction of traits in
resolution (#8205) and additional behaviour from the experimental prune
dependencies flag, this check should now pass as expected.

### Modifications:

Re-enabled the `XCTAssertFalse(stderr.contains("warning:"))` check in
the `TraitTests` where necessary and propagated the `pruneDependencies`
flag to the `ManifestLoader` init in areas it was previously missing to
assure no warnings would be generated upon omitting them. Also re-added
the enabled traits parameter done during a computation for required
dependencies when creating the package builders in
`ModulesGraph+Loading.swift`.

### Result:

This check shouldn't fail anymore.