PluginAPI: Optionally animate build progress
When a package plugin asks SwiftPM to build a target, SwiftPM captures
the whole build output including the progress messages and compiler
diagnostics into a non-TTY output buffer stream. Use of the non-TTY
output buffer stream makes the `[n/N] Compiling ...` progress messages
always to be newlined. This is not ideal when building a large scale
project as it consumes a lot of space in the terminal buffer.

This commit introduces a new parameter `progressToConsole` to the
`BuildParameters` to allow the package plugin to decide whether to
write the progress messages to the console or capture them into
`logText` buffer.