Allow creation of non-existant directory during `swift package init`
Running `swift package init --package-path <dir-does-not-exists>`
results in an error when SPM attempts to change directories in to the
directory that doesn't exist.

Add a new boolean to the `_SwiftCommand` protocol that lets commands opt
in to creating the directory at `--package-path` if it doesn't exist.
Opt in `InitCommand` to this behaviour.

Issue: #8393