Add a `defaultIsolation` static method on `SwiftSetting`. (#8301)
Add a `defaultIsolation` static method on `SwiftSetting` to allow
specifying default actor isolation within a module.

### Motivation:

This is part of
https://forums.swift.org/t/pitch-control-default-actor-isolation-inference/77482.
Please leave feedback on the API shape in the pitch thread.

### Modifications:

I added a `defaultIsolation` case to `TargetBuildSettingDescription` and
a corresponding static method to `SwiftSetting`.

### Result:

Programmers will be allowed to set `defaultIsolation` per target in a
package manifest, e.g.:

```swift
swiftSettings: [
  .enableExperimentalFeature("StrictConcurrency"),
  .defaultIsolation(MainActor.self),
]
```

---------

Co-authored-by: Pavel Yaskevich <pyaskevich@apple.com>