Add testing trait to indicate a CI environment that has restricted access.
In some testing environments, network access is limited not allowing
for hosts to access the public internet.  Add a testing condition trait
that can be used to determine if the test has unrestricted network access.

Environments that have this constraint can set `SWIFTCI_RESTRICTED_NETWORK_ACCESS=1`
to affect this condition.