[Traits] Fix traits omitting used package dependencies (#8399)
Fixes #8398 

Used package dependencies of a manifest were considered trait-guarded if
a target dependency in the manifest:

1. used a product from said package and
2. was trait-guarded

regardless as to whether the same package dependency was used elsewhere
(whether in the same target or another in the same manifest) and was not
guarded by traits.

This change ensures that this is considered when doing the calculations
for package dependencies that should be omitted if *every* target
dependency that uses said package dependency is guarded.