Tests: Skip failing tests on windows (#8210)
The Pipelines do not currently test on windows platform. To get to a
point where we can execute tests on windows, we will first disable the
failing tests and later enable them as the pipeline is setup to ensure
we do not regress in behaviour.

related to #8121
rdar://139977454


### Motivation:

In preparation for #8176 , disable failing Windows tests.

### Modifications:

Mark some tests as Skipped on Windows

Fixes #8176