Move files when using `package add-target` on single target package (#8413)
### Motivation:

Currently running `swift package init --type executable` in a folder
called `Example` creates a project with the following structure:

```
./Example/Sources
./Example/Sources/main.swift
./Example/Package.swift
```

Following this up with a `swift package add-target Foo` produces a
package that no longer builds. It now has the structure:

```
./Example/Sources
./Example/Sources/main.swift
./Example/Sources/Foo/Foo.swift
./Example/Package.swift
```

Packages with multiple targets cannot have code directly in `./Sources`,
and the user gets the error:

`Source files for target Example should be located under
'Sources/Example'`

### Modifications:

To work around this in the AddTarget command we can check if a package
is structured as a single target package with sources directly in
`./Sources` and move these files into a new folder located at
`./Sources/Example`.

### Result:

This allows the project to build after the new target has been added.

Issue: #8410

---------

Co-authored-by: Max Desiatov <m_desiatov@apple.com>