Fix typo, refine wording in `Usage.md` (#8423)
`Avoiding specifying` -> `Avoid specifying`

Also reordered words in this somewhat convoluted sentence. "provided by the library in the `module.modulemap`" could've been read as if the library provides something in such modulemap file. But in fact, "in the `module.modulemap`" is related to the first part of the sentence. Moving it there makes the sentence less ambiguous.