Fix prebuild commands
Somewhere along the way earling in the 6.1 branch changes were
made to how the prebuild commands did their check to make sure the
executable invoked in the command existed in the scratch dir.

It had always used a trick that kept track of whether a tool was
a vendored binaryTarget or built. And only succeed if it was
vendored. That information was lost in the restructuring.

This changes the check to simply check if the file exists. This
check is done after the binary artifact has been extracted so it
should be there.