Tighten up how the executable is checked.
We really want to make sure the tool isn't a built tool since it
won't always be available whent the prebuild command runs.
Instead of checking file exists, save away the source of the
plugin tool, built or vended, and pass in the paths to the built
tools to be checked in the delegate.