Sever TSC/Driver dependencies if using SwiftBuild Framework
The SwiftBuild project has a dependency on Swift Driver, which has a
dependency on Swift Tools Support Core (STSC). Swift Package Manager
(SwiftPM) executable targets also have a dependency on STSC.

Sever SwiftPM package dependency on STSC if the
SWIFTPM_SWBUILD_FRAMEWORK environment variable is set so we can
indirectly pull the STCS dependency on the dynamic library that will be
pulled in via SwiftDriver.