Bootstrap: update logic to support pre-and post python 3.8
PR #8373 (commit ID: 0193d5fb8) updated a shutil.copytree(...) command
to pass in the `dirs_exists_ok=True` argument.  However, Python added
support for the `dirs_exists_ok` argument to `shutil.copytree` in 3.8.

Update the function call to inspect the python version, and only add
the argument if the script is running in a Python version equal, or
higher than 3.8.

rdar://147297864